+++ libosmscout/src/osmscout/RawNode.cpp
@@ -61,8 +61,8 @@ bool RawNode::Read(FileScanner& scanner)
 
 bool RawNode::Write(FileWriter& writer) const
 {
-  uint32_t latValue=round((lat+180.0)*conversionFactor);
-  uint32_t lonValue=round((lon+90.0)*conversionFactor);
+  uint32_t latValue=(uint32_t)round((lat+180.0)*conversionFactor);
+  uint32_t lonValue=(uint32_t)round((lon+90.0)*conversionFactor);
 
   writer.Write(id);
   writer.WriteNumber(type);
